@@ -1,613 +1,612 @@
 cordova.define('cordova/plugin_list', function(require, exports, module) {
 module.exports = [
-    {
-        "id": "cordova-plugin-battery-status.battery",
-        "file": "plugins/cordova-plugin-battery-status/www/battery.js",
-        "pluginId": "cordova-plugin-battery-status",
-        "clobbers": [
-            "navigator.battery"
-        ]
-    },
-    {
-        "id": "cordova-plugin-camera.Camera",
-        "file": "plugins/cordova-plugin-camera/www/CameraConstants.js",
-        "pluginId": "cordova-plugin-camera",
-        "clobbers": [
-            "Camera"
-        ]
-    },
-    {
-        "id": "cordova-plugin-camera.CameraPopoverOptions",
-        "file": "plugins/cordova-plugin-camera/www/CameraPopoverOptions.js",
-        "pluginId": "cordova-plugin-camera",
-        "clobbers": [
-            "CameraPopoverOptions"
-        ]
-    },
-    {
-        "id": "cordova-plugin-camera.camera",
-        "file": "plugins/cordova-plugin-camera/www/Camera.js",
-        "pluginId": "cordova-plugin-camera",
-        "clobbers": [
-            "navigator.camera"
-        ]
-    },
-    {
-        "id": "cordova-plugin-camera.CameraPopoverHandle",
-        "file": "plugins/cordova-plugin-camera/www/CameraPopoverHandle.js",
-        "pluginId": "cordova-plugin-camera",
-        "clobbers": [
-            "CameraPopoverHandle"
-        ]
-    },
-    {
-        "id": "cordova-plugin-device.device",
-        "file": "plugins/cordova-plugin-device/www/device.js",
-        "pluginId": "cordova-plugin-device",
-        "clobbers": [
-            "device"
-        ]
-    },
-    {
-        "id": "cordova-plugin-dialogs.notification",
-        "file": "plugins/cordova-plugin-dialogs/www/notification.js",
-        "pluginId": "cordova-plugin-dialogs",
-        "merges": [
-            "navigator.notification"
-        ]
-    },
-    {
-        "id": "cordova-plugin-dialogs.notification_android",
-        "file": "plugins/cordova-plugin-dialogs/www/android/notification.js",
-        "pluginId": "cordova-plugin-dialogs",
-        "merges": [
-            "navigator.notification"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.DirectoryEntry",
-        "file": "plugins/cordova-plugin-file/www/DirectoryEntry.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.DirectoryEntry"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.DirectoryReader",
-        "file": "plugins/cordova-plugin-file/www/DirectoryReader.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.DirectoryReader"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.Entry",
-        "file": "plugins/cordova-plugin-file/www/Entry.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.Entry"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.File",
-        "file": "plugins/cordova-plugin-file/www/File.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.File"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.FileEntry",
-        "file": "plugins/cordova-plugin-file/www/FileEntry.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.FileEntry"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.FileError",
-        "file": "plugins/cordova-plugin-file/www/FileError.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.FileError"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.FileReader",
-        "file": "plugins/cordova-plugin-file/www/FileReader.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.FileReader"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.FileSystem",
-        "file": "plugins/cordova-plugin-file/www/FileSystem.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.FileSystem"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.FileUploadOptions",
-        "file": "plugins/cordova-plugin-file/www/FileUploadOptions.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.FileUploadOptions"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.FileUploadResult",
-        "file": "plugins/cordova-plugin-file/www/FileUploadResult.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.FileUploadResult"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.FileWriter",
-        "file": "plugins/cordova-plugin-file/www/FileWriter.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.FileWriter"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.Flags",
-        "file": "plugins/cordova-plugin-file/www/Flags.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.Flags"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.LocalFileSystem",
-        "file": "plugins/cordova-plugin-file/www/LocalFileSystem.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.LocalFileSystem"
-        ],
-        "merges": [
-            "window"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.Metadata",
-        "file": "plugins/cordova-plugin-file/www/Metadata.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.Metadata"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.ProgressEvent",
-        "file": "plugins/cordova-plugin-file/www/ProgressEvent.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.ProgressEvent"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.fileSystems",
-        "file": "plugins/cordova-plugin-file/www/fileSystems.js",
-        "pluginId": "cordova-plugin-file"
-    },
-    {
-        "id": "cordova-plugin-file.requestFileSystem",
-        "file": "plugins/cordova-plugin-file/www/requestFileSystem.js",
-        "pluginId": "cordova-plugin-file",
-        "clobbers": [
-            "window.requestFileSystem"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.resolveLocalFileSystemURI",
-        "file": "plugins/cordova-plugin-file/www/resolveLocalFileSystemURI.js",
-        "pluginId": "cordova-plugin-file",
-        "merges": [
-            "window"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.isChrome",
-        "file": "plugins/cordova-plugin-file/www/browser/isChrome.js",
-        "pluginId": "cordova-plugin-file",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-file.androidFileSystem",
-        "file": "plugins/cordova-plugin-file/www/android/FileSystem.js",
-        "pluginId": "cordova-plugin-file",
-        "merges": [
-            "FileSystem"
-        ]
-    },
-    {
-        "id": "cordova-plugin-file.fileSystems-roots",
-        "file": "plugins/cordova-plugin-file/www/fileSystems-roots.js",
-        "pluginId": "cordova-plugin-file",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-file.fileSystemPaths",
-        "file": "plugins/cordova-plugin-file/www/fileSystemPaths.js",
-        "pluginId": "cordova-plugin-file",
-        "merges": [
-            "cordova"
-        ],
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-geolocation.geolocation",
-        "file": "plugins/cordova-plugin-geolocation/www/android/geolocation.js",
-        "pluginId": "cordova-plugin-geolocation",
-        "clobbers": [
-            "navigator.geolocation"
-        ]
-    },
-    {
-        "id": "cordova-plugin-geolocation.PositionError",
-        "file": "plugins/cordova-plugin-geolocation/www/PositionError.js",
-        "pluginId": "cordova-plugin-geolocation",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-globalization.GlobalizationError",
-        "file": "plugins/cordova-plugin-globalization/www/GlobalizationError.js",
-        "pluginId": "cordova-plugin-globalization",
-        "clobbers": [
-            "window.GlobalizationError"
-        ]
-    },
-    {
-        "id": "cordova-plugin-globalization.globalization",
-        "file": "plugins/cordova-plugin-globalization/www/globalization.js",
-        "pluginId": "cordova-plugin-globalization",
-        "clobbers": [
-            "navigator.globalization"
-        ]
-    },
-    {
-        "id": "cordova-plugin-inappbrowser.inappbrowser",
-        "file": "plugins/cordova-plugin-inappbrowser/www/inappbrowser.js",
-        "pluginId": "cordova-plugin-inappbrowser",
-        "clobbers": [
-            "cordova.InAppBrowser.open",
-            "window.open"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.CaptureAudioOptions",
-        "file": "plugins/cordova-plugin-media-capture/www/CaptureAudioOptions.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "clobbers": [
-            "CaptureAudioOptions"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.CaptureImageOptions",
-        "file": "plugins/cordova-plugin-media-capture/www/CaptureImageOptions.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "clobbers": [
-            "CaptureImageOptions"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.CaptureVideoOptions",
-        "file": "plugins/cordova-plugin-media-capture/www/CaptureVideoOptions.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "clobbers": [
-            "CaptureVideoOptions"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.CaptureError",
-        "file": "plugins/cordova-plugin-media-capture/www/CaptureError.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "clobbers": [
-            "CaptureError"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.MediaFileData",
-        "file": "plugins/cordova-plugin-media-capture/www/MediaFileData.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "clobbers": [
-            "MediaFileData"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.MediaFile",
-        "file": "plugins/cordova-plugin-media-capture/www/MediaFile.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "clobbers": [
-            "MediaFile"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.helpers",
-        "file": "plugins/cordova-plugin-media-capture/www/helpers.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-media-capture.capture",
-        "file": "plugins/cordova-plugin-media-capture/www/capture.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "clobbers": [
-            "navigator.device.capture"
-        ]
-    },
-    {
-        "id": "cordova-plugin-media-capture.init",
-        "file": "plugins/cordova-plugin-media-capture/www/android/init.js",
-        "pluginId": "cordova-plugin-media-capture",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-network-information.network",
-        "file": "plugins/cordova-plugin-network-information/www/network.js",
-        "pluginId": "cordova-plugin-network-information",
-        "clobbers": [
-            "navigator.connection",
-            "navigator.network.connection"
-        ]
-    },
-    {
-        "id": "cordova-plugin-network-information.Connection",
-        "file": "plugins/cordova-plugin-network-information/www/Connection.js",
-        "pluginId": "cordova-plugin-network-information",
-        "clobbers": [
-            "Connection"
-        ]
-    },
-    {
-        "id": "cordova-plugin-vibration.notification",
-        "file": "plugins/cordova-plugin-vibration/www/vibration.js",
-        "pluginId": "cordova-plugin-vibration",
-        "merges": [
-            "navigator.notification",
-            "navigator"
-        ]
-    },
-    {
-        "id": "phonegap-plugin-barcodescanner.BarcodeScanner",
-        "file": "plugins/phonegap-plugin-barcodescanner/www/barcodescanner.js",
-        "pluginId": "phonegap-plugin-barcodescanner",
-        "clobbers": [
-            "cordova.plugins.barcodeScanner"
-        ]
-    },
-    {
-        "id": "cordova-plugin-googlemaps.BaseClass",
-        "file": "plugins/cordova-plugin-googlemaps/www/BaseClass.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.BaseArrayClass",
-        "file": "plugins/cordova-plugin-googlemaps/www/BaseArrayClass.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.LatLng",
-        "file": "plugins/cordova-plugin-googlemaps/www/LatLng.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.LatLngBounds",
-        "file": "plugins/cordova-plugin-googlemaps/www/LatLngBounds.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.VisibleRegion",
-        "file": "plugins/cordova-plugin-googlemaps/www/VisibleRegion.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Location",
-        "file": "plugins/cordova-plugin-googlemaps/www/Location.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.CameraPosition",
-        "file": "plugins/cordova-plugin-googlemaps/www/CameraPosition.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Polyline",
-        "file": "plugins/cordova-plugin-googlemaps/www/Polyline.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Polygon",
-        "file": "plugins/cordova-plugin-googlemaps/www/Polygon.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Marker",
-        "file": "plugins/cordova-plugin-googlemaps/www/Marker.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.HtmlInfoWindow",
-        "file": "plugins/cordova-plugin-googlemaps/www/HtmlInfoWindow.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Circle",
-        "file": "plugins/cordova-plugin-googlemaps/www/Circle.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.TileOverlay",
-        "file": "plugins/cordova-plugin-googlemaps/www/TileOverlay.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.GroundOverlay",
-        "file": "plugins/cordova-plugin-googlemaps/www/GroundOverlay.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Common",
-        "file": "plugins/cordova-plugin-googlemaps/www/Common.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.encoding",
-        "file": "plugins/cordova-plugin-googlemaps/www/encoding.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.spherical",
-        "file": "plugins/cordova-plugin-googlemaps/www/spherical.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.poly",
-        "file": "plugins/cordova-plugin-googlemaps/www/poly.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Geocoder",
-        "file": "plugins/cordova-plugin-googlemaps/www/Geocoder.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.LocationService",
-        "file": "plugins/cordova-plugin-googlemaps/www/LocationService.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Map",
-        "file": "plugins/cordova-plugin-googlemaps/www/Map.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.event",
-        "file": "plugins/cordova-plugin-googlemaps/www/event.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.MapTypeId",
-        "file": "plugins/cordova-plugin-googlemaps/www/MapTypeId.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.KmlOverlay",
-        "file": "plugins/cordova-plugin-googlemaps/www/KmlOverlay.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.KmlLoader",
-        "file": "plugins/cordova-plugin-googlemaps/www/KmlLoader.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Environment",
-        "file": "plugins/cordova-plugin-googlemaps/www/Environment.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.MarkerCluster",
-        "file": "plugins/cordova-plugin-googlemaps/www/MarkerCluster.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Cluster",
-        "file": "plugins/cordova-plugin-googlemaps/www/Cluster.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.geomodel",
-        "file": "plugins/cordova-plugin-googlemaps/www/geomodel.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.commandQueueExecutor",
-        "file": "plugins/cordova-plugin-googlemaps/www/commandQueueExecutor.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.pluginInit",
-        "file": "plugins/cordova-plugin-googlemaps/www/pluginInit.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.CordovaGoogleMaps",
-        "file": "plugins/cordova-plugin-googlemaps/www/CordovaGoogleMaps.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.StreetViewPanorama",
-        "file": "plugins/cordova-plugin-googlemaps/www/StreetViewPanorama.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.Overlay",
-        "file": "plugins/cordova-plugin-googlemaps/www/Overlay.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "runs": true
-    },
-    {
-        "id": "cordova-plugin-googlemaps.googlemaps-cdv-plugin",
-        "file": "plugins/cordova-plugin-googlemaps/www/googlemaps-cdv-plugin.js",
-        "pluginId": "cordova-plugin-googlemaps",
-        "clobbers": [
-            "plugin.google.maps"
-        ]
-    }
+  {
+    "id": "cordova-plugin-battery-status.battery",
+    "file": "plugins/cordova-plugin-battery-status/www/battery.js",
+    "pluginId": "cordova-plugin-battery-status",
+    "clobbers": [
+      "navigator.battery"
+    ]
+  },
+  {
+    "id": "cordova-plugin-camera.Camera",
+    "file": "plugins/cordova-plugin-camera/www/CameraConstants.js",
+    "pluginId": "cordova-plugin-camera",
+    "clobbers": [
+      "Camera"
+    ]
+  },
+  {
+    "id": "cordova-plugin-camera.CameraPopoverOptions",
+    "file": "plugins/cordova-plugin-camera/www/CameraPopoverOptions.js",
+    "pluginId": "cordova-plugin-camera",
+    "clobbers": [
+      "CameraPopoverOptions"
+    ]
+  },
+  {
+    "id": "cordova-plugin-camera.camera",
+    "file": "plugins/cordova-plugin-camera/www/Camera.js",
+    "pluginId": "cordova-plugin-camera",
+    "clobbers": [
+      "navigator.camera"
+    ]
+  },
+  {
+    "id": "cordova-plugin-camera.CameraPopoverHandle",
+    "file": "plugins/cordova-plugin-camera/www/CameraPopoverHandle.js",
+    "pluginId": "cordova-plugin-camera",
+    "clobbers": [
+      "CameraPopoverHandle"
+    ]
+  },
+  {
+    "id": "cordova-plugin-device.device",
+    "file": "plugins/cordova-plugin-device/www/device.js",
+    "pluginId": "cordova-plugin-device",
+    "clobbers": [
+      "device"
+    ]
+  },
+  {
+    "id": "cordova-plugin-dialogs.notification",
+    "file": "plugins/cordova-plugin-dialogs/www/notification.js",
+    "pluginId": "cordova-plugin-dialogs",
+    "merges": [
+      "navigator.notification"
+    ]
+  },
+  {
+    "id": "cordova-plugin-dialogs.notification_android",
+    "file": "plugins/cordova-plugin-dialogs/www/android/notification.js",
+    "pluginId": "cordova-plugin-dialogs",
+    "merges": [
+      "navigator.notification"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.DirectoryEntry",
+    "file": "plugins/cordova-plugin-file/www/DirectoryEntry.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.DirectoryEntry"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.DirectoryReader",
+    "file": "plugins/cordova-plugin-file/www/DirectoryReader.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.DirectoryReader"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.Entry",
+    "file": "plugins/cordova-plugin-file/www/Entry.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.Entry"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.File",
+    "file": "plugins/cordova-plugin-file/www/File.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.File"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileEntry",
+    "file": "plugins/cordova-plugin-file/www/FileEntry.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileEntry"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileError",
+    "file": "plugins/cordova-plugin-file/www/FileError.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileError"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileReader",
+    "file": "plugins/cordova-plugin-file/www/FileReader.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileReader"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileSystem",
+    "file": "plugins/cordova-plugin-file/www/FileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileSystem"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileUploadOptions",
+    "file": "plugins/cordova-plugin-file/www/FileUploadOptions.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileUploadOptions"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileUploadResult",
+    "file": "plugins/cordova-plugin-file/www/FileUploadResult.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileUploadResult"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileWriter",
+    "file": "plugins/cordova-plugin-file/www/FileWriter.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileWriter"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.Flags",
+    "file": "plugins/cordova-plugin-file/www/Flags.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.Flags"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.LocalFileSystem",
+    "file": "plugins/cordova-plugin-file/www/LocalFileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.LocalFileSystem"
+    ],
+    "merges": [
+      "window"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.Metadata",
+    "file": "plugins/cordova-plugin-file/www/Metadata.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.Metadata"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.ProgressEvent",
+    "file": "plugins/cordova-plugin-file/www/ProgressEvent.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.ProgressEvent"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.fileSystems",
+    "file": "plugins/cordova-plugin-file/www/fileSystems.js",
+    "pluginId": "cordova-plugin-file"
+  },
+  {
+    "id": "cordova-plugin-file.requestFileSystem",
+    "file": "plugins/cordova-plugin-file/www/requestFileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.requestFileSystem"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.resolveLocalFileSystemURI",
+    "file": "plugins/cordova-plugin-file/www/resolveLocalFileSystemURI.js",
+    "pluginId": "cordova-plugin-file",
+    "merges": [
+      "window"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.isChrome",
+    "file": "plugins/cordova-plugin-file/www/browser/isChrome.js",
+    "pluginId": "cordova-plugin-file",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-file.androidFileSystem",
+    "file": "plugins/cordova-plugin-file/www/android/FileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "merges": [
+      "FileSystem"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.fileSystems-roots",
+    "file": "plugins/cordova-plugin-file/www/fileSystems-roots.js",
+    "pluginId": "cordova-plugin-file",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-file.fileSystemPaths",
+    "file": "plugins/cordova-plugin-file/www/fileSystemPaths.js",
+    "pluginId": "cordova-plugin-file",
+    "merges": [
+      "cordova"
+    ],
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-geolocation.geolocation",
+    "file": "plugins/cordova-plugin-geolocation/www/android/geolocation.js",
+    "pluginId": "cordova-plugin-geolocation",
+    "clobbers": [
+      "navigator.geolocation"
+    ]
+  },
+  {
+    "id": "cordova-plugin-geolocation.PositionError",
+    "file": "plugins/cordova-plugin-geolocation/www/PositionError.js",
+    "pluginId": "cordova-plugin-geolocation",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-globalization.GlobalizationError",
+    "file": "plugins/cordova-plugin-globalization/www/GlobalizationError.js",
+    "pluginId": "cordova-plugin-globalization",
+    "clobbers": [
+      "window.GlobalizationError"
+    ]
+  },
+  {
+    "id": "cordova-plugin-globalization.globalization",
+    "file": "plugins/cordova-plugin-globalization/www/globalization.js",
+    "pluginId": "cordova-plugin-globalization",
+    "clobbers": [
+      "navigator.globalization"
+    ]
+  },
+  {
+    "id": "cordova-plugin-inappbrowser.inappbrowser",
+    "file": "plugins/cordova-plugin-inappbrowser/www/inappbrowser.js",
+    "pluginId": "cordova-plugin-inappbrowser",
+    "clobbers": [
+      "cordova.InAppBrowser.open",
+      "window.open"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.CaptureAudioOptions",
+    "file": "plugins/cordova-plugin-media-capture/www/CaptureAudioOptions.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "clobbers": [
+      "CaptureAudioOptions"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.CaptureImageOptions",
+    "file": "plugins/cordova-plugin-media-capture/www/CaptureImageOptions.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "clobbers": [
+      "CaptureImageOptions"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.CaptureVideoOptions",
+    "file": "plugins/cordova-plugin-media-capture/www/CaptureVideoOptions.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "clobbers": [
+      "CaptureVideoOptions"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.CaptureError",
+    "file": "plugins/cordova-plugin-media-capture/www/CaptureError.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "clobbers": [
+      "CaptureError"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.MediaFileData",
+    "file": "plugins/cordova-plugin-media-capture/www/MediaFileData.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "clobbers": [
+      "MediaFileData"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.MediaFile",
+    "file": "plugins/cordova-plugin-media-capture/www/MediaFile.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "clobbers": [
+      "MediaFile"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.helpers",
+    "file": "plugins/cordova-plugin-media-capture/www/helpers.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-media-capture.capture",
+    "file": "plugins/cordova-plugin-media-capture/www/capture.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "clobbers": [
+      "navigator.device.capture"
+    ]
+  },
+  {
+    "id": "cordova-plugin-media-capture.init",
+    "file": "plugins/cordova-plugin-media-capture/www/android/init.js",
+    "pluginId": "cordova-plugin-media-capture",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-network-information.network",
+    "file": "plugins/cordova-plugin-network-information/www/network.js",
+    "pluginId": "cordova-plugin-network-information",
+    "clobbers": [
+      "navigator.connection",
+      "navigator.network.connection"
+    ]
+  },
+  {
+    "id": "cordova-plugin-network-information.Connection",
+    "file": "plugins/cordova-plugin-network-information/www/Connection.js",
+    "pluginId": "cordova-plugin-network-information",
+    "clobbers": [
+      "Connection"
+    ]
+  },
+  {
+    "id": "cordova-plugin-vibration.notification",
+    "file": "plugins/cordova-plugin-vibration/www/vibration.js",
+    "pluginId": "cordova-plugin-vibration",
+    "merges": [
+      "navigator.notification",
+      "navigator"
+    ]
+  },
+  {
+    "id": "phonegap-plugin-barcodescanner.BarcodeScanner",
+    "file": "plugins/phonegap-plugin-barcodescanner/www/barcodescanner.js",
+    "pluginId": "phonegap-plugin-barcodescanner",
+    "clobbers": [
+      "cordova.plugins.barcodeScanner"
+    ]
+  },
+  {
+    "id": "cordova-plugin-googlemaps.BaseClass",
+    "file": "plugins/cordova-plugin-googlemaps/www/BaseClass.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.BaseArrayClass",
+    "file": "plugins/cordova-plugin-googlemaps/www/BaseArrayClass.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.LatLng",
+    "file": "plugins/cordova-plugin-googlemaps/www/LatLng.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.LatLngBounds",
+    "file": "plugins/cordova-plugin-googlemaps/www/LatLngBounds.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.VisibleRegion",
+    "file": "plugins/cordova-plugin-googlemaps/www/VisibleRegion.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Location",
+    "file": "plugins/cordova-plugin-googlemaps/www/Location.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.CameraPosition",
+    "file": "plugins/cordova-plugin-googlemaps/www/CameraPosition.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Polyline",
+    "file": "plugins/cordova-plugin-googlemaps/www/Polyline.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Polygon",
+    "file": "plugins/cordova-plugin-googlemaps/www/Polygon.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Marker",
+    "file": "plugins/cordova-plugin-googlemaps/www/Marker.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.HtmlInfoWindow",
+    "file": "plugins/cordova-plugin-googlemaps/www/HtmlInfoWindow.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Circle",
+    "file": "plugins/cordova-plugin-googlemaps/www/Circle.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.TileOverlay",
+    "file": "plugins/cordova-plugin-googlemaps/www/TileOverlay.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.GroundOverlay",
+    "file": "plugins/cordova-plugin-googlemaps/www/GroundOverlay.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Common",
+    "file": "plugins/cordova-plugin-googlemaps/www/Common.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.encoding",
+    "file": "plugins/cordova-plugin-googlemaps/www/encoding.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.spherical",
+    "file": "plugins/cordova-plugin-googlemaps/www/spherical.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.poly",
+    "file": "plugins/cordova-plugin-googlemaps/www/poly.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Geocoder",
+    "file": "plugins/cordova-plugin-googlemaps/www/Geocoder.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.LocationService",
+    "file": "plugins/cordova-plugin-googlemaps/www/LocationService.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Map",
+    "file": "plugins/cordova-plugin-googlemaps/www/Map.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.event",
+    "file": "plugins/cordova-plugin-googlemaps/www/event.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.MapTypeId",
+    "file": "plugins/cordova-plugin-googlemaps/www/MapTypeId.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.KmlOverlay",
+    "file": "plugins/cordova-plugin-googlemaps/www/KmlOverlay.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.KmlLoader",
+    "file": "plugins/cordova-plugin-googlemaps/www/KmlLoader.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Environment",
+    "file": "plugins/cordova-plugin-googlemaps/www/Environment.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.MarkerCluster",
+    "file": "plugins/cordova-plugin-googlemaps/www/MarkerCluster.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Cluster",
+    "file": "plugins/cordova-plugin-googlemaps/www/Cluster.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.geomodel",
+    "file": "plugins/cordova-plugin-googlemaps/www/geomodel.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.commandQueueExecutor",
+    "file": "plugins/cordova-plugin-googlemaps/www/commandQueueExecutor.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.pluginInit",
+    "file": "plugins/cordova-plugin-googlemaps/www/pluginInit.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.CordovaGoogleMaps",
+    "file": "plugins/cordova-plugin-googlemaps/www/CordovaGoogleMaps.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.StreetViewPanorama",
+    "file": "plugins/cordova-plugin-googlemaps/www/StreetViewPanorama.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.Overlay",
+    "file": "plugins/cordova-plugin-googlemaps/www/Overlay.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-googlemaps.googlemaps-cdv-plugin",
+    "file": "plugins/cordova-plugin-googlemaps/www/googlemaps-cdv-plugin.js",
+    "pluginId": "cordova-plugin-googlemaps",
+    "clobbers": [
+      "plugin.google.maps"
+    ]
+  }
 ];
 module.exports.metadata = 
 // TOP OF METADATA
 {
-    "cordova-plugin-battery-status": "1.2.5",
-    "cordova-plugin-compat": "1.2.0",
-    "cordova-plugin-camera": "2.4.1",
-    "cordova-plugin-device": "1.1.7",
-    "cordova-plugin-dialogs": "1.3.4",
-    "cordova-plugin-file": "4.3.3",
-    "cordova-plugin-geolocation": "2.4.3",
-    "cordova-plugin-globalization": "1.0.9",
-    "cordova-plugin-inappbrowser": "1.7.2",
-    "cordova-plugin-media-capture": "1.4.3",
-    "cordova-plugin-network-information": "1.3.4",
-    "cordova-plugin-vibration": "2.1.6",
-    "phonegap-plugin-barcodescanner": "4.0.1",
-    "cordova-plugin-googlemaps": "2.3.8"
+  "cordova-plugin-battery-status": "1.2.5",
+  "cordova-plugin-camera": "2.4.1",
+  "cordova-plugin-device": "1.1.7",
+  "cordova-plugin-dialogs": "1.3.4",
+  "cordova-plugin-file": "4.3.3",
+  "cordova-plugin-geolocation": "2.4.3",
+  "cordova-plugin-globalization": "1.0.9",
+  "cordova-plugin-inappbrowser": "1.7.2",
+  "cordova-plugin-media-capture": "1.4.3",
+  "cordova-plugin-network-information": "1.3.4",
+  "cordova-plugin-vibration": "2.1.6",
+  "phonegap-plugin-barcodescanner": "8.0.0",
+  "cordova-plugin-googlemaps": "2.3.8"
 };
 // BOTTOM OF METADATA
 });
