@@ -16,83 +16,55 @@
        specific language governing permissions and limitations
        under the License.
 */
+/* eslint no-self-assign: 0 */
+/* eslint no-unused-vars: 0 */
 
 var Q = require('q');
 var fs = require('fs');
 var path = require('path');
 var shell = require('shelljs');
 var events = require('cordova-common').events;
-var CordovaError = require('cordova-common').CordovaError;
 
 function GenericBuilder (projectDir) {
     this.root = projectDir || path.resolve(__dirname, '../../..');
     this.binDirs = {
-        ant: path.join(this.root, hasCustomRules(this.root) ? 'ant-build' : 'bin'),
+        studio: path.join(this.root, 'app', 'build', 'outputs', 'apk'),
         gradle: path.join(this.root, 'build', 'outputs', 'apk')
     };
 }
 
-function hasCustomRules(projectRoot) {
-    return fs.existsSync(path.join(projectRoot, 'custom_rules.xml'));
-}
-
-GenericBuilder.prototype.prepEnv = function() {
+GenericBuilder.prototype.prepEnv = function () {
     return Q();
 };
 
-GenericBuilder.prototype.build = function() {
+GenericBuilder.prototype.build = function () {
     events.emit('log', 'Skipping build...');
     return Q(null);
 };
 
-GenericBuilder.prototype.clean = function() {
+GenericBuilder.prototype.clean = function () {
     return Q();
 };
 
-GenericBuilder.prototype.findOutputApks = function(build_type, arch) {
+GenericBuilder.prototype.findOutputApks = function (build_type, arch) {
     var self = this;
-    return Object.keys(this.binDirs)
-    .reduce(function (result, builderName) {
+    return Object.keys(this.binDirs).reduce(function (result, builderName) {
         var binDir = self.binDirs[builderName];
         return result.concat(findOutputApksHelper(binDir, build_type, builderName === 'ant' ? null : arch));
-    }, [])
-    .sort(apkSorter);
-};
-
-GenericBuilder.prototype.readProjectProperties = function () {
-    function findAllUniq(data, r) {
-        var s = {};
-        var m;
-        while ((m = r.exec(data))) {
-            s[m[1]] = 1;
-        }
-        return Object.keys(s);
-    }
-
-    var data = fs.readFileSync(path.join(this.root, 'project.properties'), 'utf8');
-    return {
-        libs: findAllUniq(data, /^\s*android\.library\.reference\.\d+=(.*)(?:\s|$)/mg),
-        gradleIncludes: findAllUniq(data, /^\s*cordova\.gradle\.include\.\d+=(.*)(?:\s|$)/mg),
-        systemLibs: findAllUniq(data, /^\s*cordova\.system\.library\.\d+=(.*)(?:\s|$)/mg)
-    };
-};
-
-GenericBuilder.prototype.extractRealProjectNameFromManifest = function () {
-    var manifestPath = path.join(this.root, 'AndroidManifest.xml');
-    var manifestData = fs.readFileSync(manifestPath, 'utf8');
-    var m = /<manifest[\s\S]*?package\s*=\s*"(.*?)"/i.exec(manifestData);
-    if (!m) {
-        throw new CordovaError('Could not find package name in ' + manifestPath);
-    }
-
-    var packageName=m[1];
-    var lastDotIndex = packageName.lastIndexOf('.');
-    return packageName.substring(lastDotIndex + 1);
+    }, []).sort(apkSorter);
 };
 
 module.exports = GenericBuilder;
 
-function apkSorter(fileA, fileB) {
+function apkSorter (fileA, fileB) {
+    // De-prioritize arch specific builds
+    var archSpecificRE = /-x86|-arm/;
+    if (archSpecificRE.exec(fileA)) {
+        return 1;
+    } else if (archSpecificRE.exec(fileB)) {
+        return -1;
+    }
+
     // De-prioritize unsigned builds
     var unsignedRE = /-unsigned/;
     if (unsignedRE.exec(fileA)) {
@@ -101,16 +73,22 @@ function apkSorter(fileA, fileB) {
         return -1;
     }
 
-    var timeDiff = fs.statSync(fileA).mtime - fs.statSync(fileB).mtime;
+    var timeDiff = fs.statSync(fileB).mtime - fs.statSync(fileA).mtime;
     return timeDiff === 0 ? fileA.length - fileB.length : timeDiff;
 }
 
-function findOutputApksHelper(dir, build_type, arch) {
+function findOutputApksHelper (dir, build_type, arch) {
     var shellSilent = shell.config.silent;
     shell.config.silent = true;
 
-    var ret = shell.ls(path.join(dir, '*.apk'))
-    .filter(function(candidate) {
+    // list directory recursively
+    var ret = shell.ls('-R', dir).map(function (file) {
+        // ls does not include base directory
+        return path.join(dir, file);
+    }).filter(function (file) {
+        // find all APKs
+        return file.match(/\.apk?$/i);
+    }).filter(function (candidate) {
         var apkName = path.basename(candidate);
         // Need to choose between release and debug .apk.
         if (build_type === 'debug') {
@@ -120,8 +98,7 @@ function findOutputApksHelper(dir, build_type, arch) {
             return /-release/.exec(apkName) && !/-unaligned/.exec(apkName);
         }
         return true;
-    })
-    .sort(apkSorter);
+    }).sort(apkSorter);
 
     shellSilent = shellSilent;
 
@@ -131,15 +108,15 @@ function findOutputApksHelper(dir, build_type, arch) {
     // Assume arch-specific build if newest apk has -x86 or -arm.
     var archSpecific = !!/-x86|-arm/.exec(path.basename(ret[0]));
     // And show only arch-specific ones (or non-arch-specific)
-    ret = ret.filter(function(p) {
-        /*jshint -W018 */
-        return !!/-x86|-arm/.exec(path.basename(p)) == archSpecific;
-        /*jshint +W018 */
+    ret = ret.filter(function (p) {
+        /* jshint -W018 */
+        return !!/-x86|-arm/.exec(path.basename(p)) === archSpecific;
+        /* jshint +W018 */
     });
 
     if (archSpecific && ret.length > 1 && arch) {
-        ret = ret.filter(function(p) {
-            return path.basename(p).indexOf('-' + arch) != -1;
+        ret = ret.filter(function (p) {
+            return path.basename(p).indexOf('-' + arch) !== -1;
         });
     }
 
